// web/packages/ra-theme/src/api/models/checkboxFormField.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { CheckboxFormFieldDefault } from "./checkboxFormFieldDefault";
import type { CheckboxFormFieldDescription } from "./checkboxFormFieldDescription";
import type { CheckboxFormFieldOptions } from "./checkboxFormFieldOptions";

export interface CheckboxFormField {
  /** Default value of the field. */
  default?: CheckboxFormFieldDefault;
  /** Description of the field. */
  description?: CheckboxFormFieldDescription;
  /** ID of the field. */
  id: string;
  options: CheckboxFormFieldOptions;
  /** Whether the field is required. */
  required: boolean;
  /** Title of the field. */
  title: string;
  /** Type of the field. */
  type?: string;
}