// web/packages/ra-theme/src/api/models/dropdownFormField.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { DropdownFormFieldDefault } from "./dropdownFormFieldDefault";
import type { DropdownFormFieldDescription } from "./dropdownFormFieldDescription";
import type { DropdownFormFieldOptions } from "./dropdownFormFieldOptions";

export interface DropdownFormField {
  /** Default value of the field. */
  default?: DropdownFormFieldDefault;
  /** Description of the field. */
  description?: DropdownFormFieldDescription;
  /** ID of the field. */
  id: string;
  options: DropdownFormFieldOptions;
  /** Whether the field is required. */
  required: boolean;
  /** Title of the field. */
  title: string;
  /** Type of the field. */
  type?: string;
}