// web/packages/ra-theme/src/api/models/emailFormField.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { EmailFormFieldDefault } from "./emailFormFieldDefault";
import type { EmailFormFieldDescription } from "./emailFormFieldDescription";

export interface EmailFormField {
  /** Default value of the field. */
  default?: EmailFormFieldDefault;
  /** Description of the field. */
  description?: EmailFormFieldDescription;
  /** ID of the field. */
  id: string;
  /** Whether the field is required. */
  required: boolean;
  /** Title of the field. */
  title: string;
  /** Type of the field. */
  type?: string;
}