// web/packages/ra-theme/src/api/models/getResponseForm.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { FormFields } from "./formFields";

/**
 * The form.
 */
export interface GetResponseForm {
  fields: FormFields;
  /** ID of the form. */
  id: string;
  /** Title of the form. */
  title: string;
}