// web/packages/ra-theme/src/api/models/numberFormField.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { NumberFormFieldDefault } from "./numberFormFieldDefault";
import type { NumberFormFieldDescription } from "./numberFormFieldDescription";

export interface NumberFormField {
  /** Default value of the field. */
  default?: NumberFormFieldDefault;
  /** Description of the field. */
  description?: NumberFormFieldDescription;
  /** ID of the field. */
  id: string;
  /** Whether the field is required. */
  required: boolean;
  /** Title of the field. */
  title: string;
  /** Type of the field. */
  type?: string;
}