// web/packages/ra-theme/src/api/models/radioFormField.ts
/**
 * Generated by orval v6.19.1 🍺
 * Do not edit manually.
 * api
 * surveys api
 * OpenAPI spec version: 0.1.0
 */
import type { RadioFormFieldDefault } from "./radioFormFieldDefault";
import type { RadioFormFieldDescription } from "./radioFormFieldDescription";
import type { RadioFormFieldOptions } from "./radioFormFieldOptions";

export interface RadioFormField {
  /** Default value of the field. */
  default?: RadioFormFieldDefault;
  /** Description of the field. */
  description?: RadioFormFieldDescription;
  /** ID of the field. */
  id: string;
  options: RadioFormFieldOptions;
  /** Whether the field is required. */
  required: boolean;
  /** Title of the field. */
  title: string;
  /** Type of the field. */
  type?: string;
}